+++ config.c
@@ -502,6 +502,9 @@ parse_section_colors(struct context *ctx)
     else if (strcmp(key, "selection-text") == 0)
         return value_to_color(ctx, true, &conf->colors.selection_text);
 
+    else if (strcmp(key, "selection-match") == 0)
+        return value_to_color(ctx, true, &conf->colors.selection_match);
+
     else if (strcmp(key, "border") == 0)
         return value_to_color(ctx, true, &conf->colors.border);
 
@@ -923,6 +926,7 @@ config_load(struct config *conf, const char *conf_path,
             .match = conf_hex_to_rgba(0xcb4b16ff),
             .selection = conf_hex_to_rgba(0xeee8d5dd),
             .selection_text = conf_hex_to_rgba(0x657b83ff),
+            .selection_match = conf_hex_to_rgba(0xcb4b16ff),
         },
         .border = {
             .size = 1u,